python 3 compatibility

+++ pyqg/model.py
@@ -1,8 +1,9 @@
+from __future__ import print_function
 import numpy as np
-from kernel import PseudoSpectralKernel, tendency_forward_euler, tendency_ab2, tendency_ab3
 from numpy import pi
 import logging
-try:   
+from .kernel import PseudoSpectralKernel, tendency_forward_euler, tendency_ab2, tendency_ab3
+try:
     import mkl
     np.use_fastnumpy = True
 except ImportError:
@@ -10,13 +11,13 @@ except ImportError:
 
 try:
     import pyfftw
-    pyfftw.interfaces.cache.enable() 
+    pyfftw.interfaces.cache.enable()
 except ImportError:
     pass
 
 class Model(PseudoSpectralKernel):
     """A generic pseudo-spectral inversion model.
-    
+
     Attributes
     ----------
     q : real array
@@ -38,8 +39,8 @@ class Model(PseudoSpectralKernel):
     rek : float
         Linear drag in lower layer
     filterfac : float
-        Amplitdue of the spectral spherical filter 
-    dt : float 
+        Amplitdue of the spectral spherical filter
+    dt : float
         Numerical timstep
     twrite : int
         Interval for cfl writeout (units: number of timesteps)
@@ -62,7 +63,7 @@ class Model(PseudoSpectralKernel):
     radii :  real array
         Deformation radii  (units: model length)
     """
-    
+
     def __init__(
         self,
         # grid size parameters
@@ -88,7 +89,7 @@ class Model(PseudoSpectralKernel):
         diagnostics_list='all',     # which diagnostics to output
         # fft parameters
         # removed because fftw is now manditory
-        #use_fftw = False,               # fftw flag 
+        #use_fftw = False,               # fftw flag
         #teststyle = False,            # use fftw with "estimate" planner to get reproducibility
         ntd = 1,                       # number of threads to use in fftw computations
         log_level = 1,                 # logger level: from 0 for quiet (no log) to 4 for verbose
@@ -99,7 +100,7 @@ class Model(PseudoSpectralKernel):
         .. note:: All of the test cases use ``nx==ny``. Expect bugs if you choose
                   these parameters to be different.
         .. note:: All time intervals will be rounded to nearest `dt` interval.
-        
+
         Parameters
         ----------
         nx : int
@@ -109,7 +110,7 @@ class Model(PseudoSpectralKernel):
         L : number
             Domain length in x direction. Units: meters.
         W :
-            Domain width in y direction. Units: meters (default: L).    
+            Domain width in y direction. Units: meters (default: L).
         rek : number
             linear drag in lower layer. Units: seconds :sup:`-1`.
         filterfac : number
@@ -130,7 +131,7 @@ class Model(PseudoSpectralKernel):
             Units: seconds. (For performance purposes, averaging does not have to
             occur every timestep)
         tsnapint : number
-            Time interval for snapshots. Units: seconds. 
+            Time interval for snapshots. Units: seconds.
         ntd : int
             Number of threads to use. Should not exceed the number of cores on
             your machine.
@@ -138,7 +139,7 @@ class Model(PseudoSpectralKernel):
 
         if ny is None: ny = nx
         if W is None: W = L
-       
+
         # put all the parameters into the object
         # grid
         self.nx = nx
@@ -155,11 +156,11 @@ class Model(PseudoSpectralKernel):
         self.logfile = logfile
         self.log_level = log_level
         self.useAB2 = useAB2
-        # fft 
+        # fft
         #self.use_fftw = use_fftw
         #self.teststyle= teststyle
         self.ntd = ntd
-        
+
         # friction
         self.rek = rek
         self.filterfac = filterfac
@@ -177,38 +178,38 @@ class Model(PseudoSpectralKernel):
         self._initialize_forcing()
         self._initialize_filter()
         self._initialize_inversion_matrix()
-        self._initialize_time()                
+        self._initialize_time()
 
         # call the underlying cython kernel
         self._initialize_kernel()
-       
+
         self._initialize_diagnostics(diagnostics_list)
-   
+
     def run_with_snapshots(self, tsnapstart=0., tsnapint=432000.):
         """Run the model forward, yielding to user code at specified intervals.
-        
+
         Parameters
         ----------
-        
+
         tsnapstart : int
             The timestep at which to begin yielding.
         tstapint : int
             The interval at which to yield.
         """
-        
+
         tsnapints = np.ceil(tsnapint/self.dt)
-        
+
         while(self.t < self.tmax):
             self._step_forward()
             if self.t>=tsnapstart and (self.tc%tsnapints)==0:
                 yield self.t
         return
-                
+
     def run(self):
         """Run the model forward without stopping until the end."""
-        while(self.t < self.tmax): 
+        while(self.t < self.tmax):
             self._step_forward()
-               
+
 
     def vertical_modes(self):
         """ Calculate standard vertical modes. Simply
@@ -224,7 +225,7 @@ class Model(PseudoSpectralKernel):
         self.radii[0] = self.g*self.H/np.abs(self.f) # barotropic def. radius
         self.radii[1:] = 1./np.sqrt(self.kdi2[1:])
 
-        # eigenstructure 
+        # eigenstructure
         self.pmodes = evecs[:,asort]
 
         # normalize to have unit L2-norm
@@ -235,7 +236,7 @@ class Model(PseudoSpectralKernel):
         """ Performs a field p into modal amplitudes pn
                 using the basis [pmodes]. The inverse
                 transform calculates p from pn"""
-        
+
         if forward:
             pt = np.linalg.solve(self.pmodes[np.newaxis,np.newaxis],p.T).T
         else:
@@ -245,7 +246,7 @@ class Model(PseudoSpectralKernel):
 
     def stability_analysis(self,bottom_friction=False):
         """ Performs the baroclinic linear instability analysis given
-                given the base state velocity :math: `(U, V)` and 
+                given the base state velocity :math: `(U, V)` and
                 the stretching matrix  :math: `S`:
 
         .. math:: A \Phi = \omega B \Phi,
@@ -256,29 +257,29 @@ class Model(PseudoSpectralKernel):
                         1j \delta_{N N} r_{ek} I  \kappa^2
 
         where :math:`\delta_{N N} = [0,0,\dots,0,1] ,`
-        
+
         and
 
         .. math:: B =  S - I \kappa^2 .
 
-        
-        The eigenstructure is 
-        
-        .. math:: \Phi 
-        
+
+        The eigenstructure is
+
+        .. math:: \Phi
+
         and the eigenvalue is
 
         .. math:: `\omega`
-        
+
         The growth rate is Im\ :math:`\{\omega\}`.
 
 
         Parameters
         ----------
         bottom_friction: optional inclusion linear bottom drag
-                         in the linear stability calculation 
+                         in the linear stability calculation
                          (default is False, as if :math: `r_{ek} = 0`)
-        
+
         Returns
         -------
         omega: complex array
@@ -310,7 +311,7 @@ class Model(PseudoSpectralKernel):
 
         evals,evecs = np.linalg.eig(M)
         evals, evecs = evals.T, evecs.T
-    
+
         # sorting things this way proved way
         #  more faster than using numpy's argsort() !
         imax = evals.imag.argmax(axis=0)
@@ -330,29 +331,29 @@ class Model(PseudoSpectralKernel):
 
         self._do_advection()
         # use streamfunction to calculate advection tendency
-        
+
         self._do_friction()
-        # apply friction 
-        
+        # apply friction
+
         self._do_external_forcing()
         # apply external forcing
-        
+
         self._calc_diagnostics()
         # do what has to be done with diagnostics
-        
+
         self._forward_timestep()
         # apply tendencies to step the model forward
         # (filter gets called here)
 
         # the basic steps are
-        self._print_status() 
-        
+        self._print_status()
+
     def _initialize_time(self):
         """Set up timestep stuff"""
         #self.t=0        # actual time
         #self.tc=0       # timestep number
-        self.taveints = np.ceil(self.taveint/self.dt)    
-        
+        self.taveints = np.ceil(self.taveint/self.dt)
+
     ### initialization routines, only called once at the beginning ###
     def _initialize_grid(self):
         """Set up spatial and spectral grids and related constants"""
@@ -367,8 +368,8 @@ class Model(PseudoSpectralKernel):
 
         # wavenumber grids
         self.nl = self.ny
-        self.nk = self.nx/2+1
-        self.ll = self.dl*np.append( np.arange(0.,self.nx/2), 
+        self.nk = int(self.nx/2+1)
+        self.ll = self.dl*np.append( np.arange(0.,self.nx/2),
             np.arange(-self.nx/2,0.) )
         self.kk = self.dk*np.arange(0.,self.nk)
 
@@ -381,24 +382,24 @@ class Model(PseudoSpectralKernel):
 
         # constant for spectral normalizations
         self.M = self.nx*self.ny
-        
+
         # isotropic wavenumber^2 grid
-        # the inversion is not defined at kappa = 0 
+        # the inversion is not defined at kappa = 0
         self.wv2 = self.k**2 + self.l**2
         self.wv = np.sqrt( self.wv2 )
 
         iwv2 = self.wv2 != 0.
         self.wv2i = np.zeros_like(self.wv2)
         self.wv2i[iwv2] = self.wv2[iwv2]**-1
-        
-    def _initialize_background(self):        
+
+    def _initialize_background(self):
         raise NotImplementedError(
             'needs to be implemented by Model subclass')
-        
+
     def _initialize_inversion_matrix(self):
         raise NotImplementedError(
             'needs to be implemented by Model subclass')
-            
+
     def _initialize_forcing(self):
         raise NotImplementedError(
             'needs to be implemented by Model subclass')
@@ -408,7 +409,7 @@ class Model(PseudoSpectralKernel):
         # this defines the spectral filter (following Arbic and Flierl, 2003)
         cphi=0.65*pi
         wvx=np.sqrt((self.k*self.dx)**2.+(self.l*self.dy)**2.)
-        self.filtr = np.exp(-self.filterfac*(wvx-cphi)**4.)  
+        self.filtr = np.exp(-self.filterfac*(wvx-cphi)**4.)
         self.filtr[wvx<=cphi] = 1.
 
     def _filter(self, q):
@@ -416,7 +417,7 @@ class Model(PseudoSpectralKernel):
 
     def _do_external_forcing(self):
         pass
-            
+
     def _initialize_kernel(self):
         #super(spectral_kernel.PseudoSpectralKernel, self).__init__(
         self._kernel_init(
@@ -428,13 +429,13 @@ class Model(PseudoSpectralKernel):
             rek=self.rek,
             fftw_num_threads=self.ntd
         )
-        
+
         # still need to initialize a few state variables here, outside kernel
         # this is sloppy
         #self.dqhdt_forc = np.zeros_like(self.qh)
         #self.dqhdt_p = np.zeros_like(self.qh)
         #self.dqhdt_pp = np.zeros_like(self.qh)
-        
+
         self.logger.info(' Kernel initialized')
 
 
@@ -457,9 +458,9 @@ class Model(PseudoSpectralKernel):
             self.logger.addHandler(fhandler)
 
         self.logger.setLevel(self.log_level*10)
-        
+
         # this prevents the logger to propagate into the ipython notebook log
-        self.logger.propagate = False   
+        self.logger.propagate = False
 
         self.logger.info(' Logger initialized')
 
@@ -473,7 +474,7 @@ class Model(PseudoSpectralKernel):
         vq = v*q
         # this is a hack, since fft now requires input to have shape (nz,ny,nx)
         # it does an extra unnecessary fft
-        is_2d = (uq.ndim==2) 
+        is_2d = (uq.ndim==2)
         if is_2d:
             uq = np.tile(uq[np.newaxis,:,:], (self.nz,1,1))
             vq = np.tile(vq[np.newaxis,:,:], (self.nz,1,1))
@@ -486,7 +487,7 @@ class Model(PseudoSpectralKernel):
     # def _filter(self, q):
     #     """Apply filter to field q."""
     #     return q
-        
+
     def _print_status(self):
         """Output some basic stats."""
         if (self.log_level) and ((self.tc % self.twrite)==0):
@@ -498,7 +499,7 @@ class Model(PseudoSpectralKernel):
                     , self.tc,self.t,self.ke,self.cfl )
 
             assert self.cfl<1., self.logger.error('CFL condition violated')
-            
+
 
     def _calc_diagnostics(self):
         # here is where we calculate diagnostics
@@ -538,46 +539,46 @@ class Model(PseudoSpectralKernel):
     def _calc_cfl(self):
         raise NotImplementedError(
             'needs to be implemented by Model subclass')
-    
+
     # this is stuff the Cesar added
-    
+
     # if self.tc==0:
     #     assert self.calc_cfl()<1., " *** time-step too large "
     #     # initialize ke and time arrays
     #     self.ke = np.array([self.calc_ke()])
     #     self.eddy_time = np.array([self.calc_eddy_time()])
     #     self.time = np.array([0.])
-           
+
 
     def _calc_ke(self):
         raise NotImplementedError(
-            'needs to be implemented by Model subclass')        
+            'needs to be implemented by Model subclass')
 
     def _initialize_diagnostics(self, diagnostics_list):
         # Initialization for diagnotics
         self.diagnostics = dict()
-        
+
         self._initialize_core_diagnostics()
         self._initialize_model_diagnostics()
-        
+
         if diagnostics_list == 'all':
             pass # by default, all diagnostics are active
         elif diagnostics_list == 'none':
             self.set_active_diagnostics([])
         else:
             self.set_active_diagnostics(diagnostics_list)
-            
+
     def _initialize_core_diagnostics(self):
         """Diagnostics common to all models."""
         self.add_diagnostic('Ensspec',
             description='enstrophy spectrum',
             function= (lambda self: np.abs(self.qh)**2/self.M**2)
         )
-        
+
         self.add_diagnostic('KEspec',
             description=' kinetic energy spectrum',
             function=(lambda self: self.wv2*np.abs(self.ph)**2/self.M**2)
-        )      # factor of 2 to account for the fact that we have only half of 
+        )      # factor of 2 to account for the fact that we have only half of
                #    the Fourier coefficients.
 
         self.add_diagnostic('q',
@@ -589,7 +590,7 @@ class Model(PseudoSpectralKernel):
             description='total energy dissipation by bottom drag',
             function= (lambda self: self.Hi[-1]/self.H*self.rek*(self.v[-1]**2 + self.u[-1]**2).mean())
         )
-        
+
         self.add_diagnostic('EKE',
             description='mean eddy kinetic energy',
             function= (lambda self: 0.5*(self.v**2 + self.u**2).mean(axis=-1).mean(axis=-1))
@@ -597,25 +598,25 @@ class Model(PseudoSpectralKernel):
 
     def _calc_derived_fields(self):
         """Should be implemented by subclass."""
-        pass        
-                 
+        pass
+
     def _initialize_model_diagnostics(self):
         """Should be implemented by subclass."""
         pass
-            
+
     def _set_active_diagnostics(self, diagnostics_list):
         for d in self.diagnostics:
             self.diagnostics[d]['active'] == (d in diagnostics_list)
-            
+
     def add_diagnostic(self, diag_name, description=None, units=None, function=None):
         # create a new diagnostic dict and add it to the object array
-        
+
         # make sure the function is callable
         assert hasattr(function, '__call__')
-        
+
         # make sure the name is valid
         assert isinstance(diag_name, str)
-        
+
         # by default, diagnostic is active
         self.diagnostics[diag_name] = {
            'description': description,
@@ -623,7 +624,7 @@ class Model(PseudoSpectralKernel):
            'active': True,
            'count': 0,
            'function': function, }
-           
+
     def describe_diagnostics(self):
         """Print a human-readable summary of the available diagnostics."""
         diag_names = self.diagnostics.keys()
@@ -634,12 +635,12 @@ class Model(PseudoSpectralKernel):
             d = self.diagnostics[k]
             print('{:<10} | {:<54}').format(
                  *(k,  d['description']))
-           
+
     def _increment_diagnostics(self):
         # compute intermediate quantities needed for some diagnostics
-        
+
         self._calc_derived_fields()
-        
+
         for dname in self.diagnostics:
             if self.diagnostics[dname]['active']:
                 res = self.diagnostics[dname]['function'](self)
@@ -648,9 +649,9 @@ class Model(PseudoSpectralKernel):
                 else:
                     self.diagnostics[dname]['value'] += res
                 self.diagnostics[dname]['count'] += 1
-                
+
     def get_diagnostic(self, dname):
-        return (self.diagnostics[dname]['value'] / 
+        return (self.diagnostics[dname]['value'] /
                 self.diagnostics[dname]['count'])
 
     def spec_var(self, ph):
@@ -660,4 +661,3 @@ class Model(PseudoSpectralKernel):
         var_dens[...,0] = var_dens[...,0]/2.
         var_dens[...,-1] = var_dens[...,-1]/2.
         return var_dens.sum()
-
